dt-bindings: sound: rt5677: Convert to json-schema
Convert the Realtek RT5677 codec bindings from the free-form text format
to json-schema.

Note that as part of this conversion, the realtek,jd{1,2,3}-gpio
properties are renamed to realtek,jd{1,2,3}-config because the -gpio
suffix conflicts with the standard GPIO consumer schema. This should be
okay to do because the only user of this binding (Google Pixel C) does
not currently set this property.

Signed-off-by: Thierry Reding <treding@nvidia.com>