ARM: tegra: Add compatible string for built-in ASIX on Colibri boards
The device tree node for the built-in ASIX Ethernet device on Colibri
boards needs a compatible string in order to pass DT schema validation.
Add the USB VID,PID compatible string as required by the DT schema for
USB devices.

Signed-off-by: Thierry Reding <treding@nvidia.com>
---
Marcel, I've taken the USB vendor and device IDs from the ASIX driver
match entry for the AX88772B module, but do you have a quick way of
verifying that that's indeed the ID that the device reports in those
systems?

Rob, I do get a checkpatch.pl warning for this new compatible string,
but I'm not sure there's anything easy that could be done about this,
other than perhaps making checkpatch.pl aware of the special USB (and
potentially PCI) compatible strings as well.