ASoC: rt5677: Update for DT schema changes
With the recent conversion of the RT5677 device tree bindings to
json-schema, the "realtek,jd{1,2,3}-gpio" properties were renamed to
"realtek,jd{1,2,3}-config" to avoid clashing with the standard GPIO
bindings. Update the driver to use these new properties.

This should be safe to do because luckily no device trees within the
kernel have been using these properties. The only occurrence of one of
them is in the example provided by the bindings.

Signed-off-by: Thierry Reding <treding@nvidia.com>