drm/nouveau: Remove unneeded gk20a_vmm_aper()
The gk20a_vmm_aper() function is no longer needed. It was created based
on the wrong assumption that apertures were somehow special in the iGPU
variants, whereas the corresponding bitfields are simply ignored and
therefore often programmed as 0.

Newer integrated GPUs such as the gv11b (found on Tegra194) perform full
checks on the aperture fields and they have to be set properly, so this
will also enable compatibility for those newer chips.

Signed-off-by: Thierry Reding <treding@nvidia.com>