driver core: platform: Add API to get optional resources
Add devm_platform_ioremap_resource_byname_optional(), which is a variant
of devm_platform_ioremap_resource_byname() that returns NULL if the
named resource cannot be found. This is useful for resources (such as
I/O memory regions) that are optional (such as privileged registers in a
virtualized version of a device).

Signed-off-by: Thierry Reding <treding@nvidia.com>